Version 1 of the schematic
After I built it referencing several other schematics on the web - I drew the schematic of what I had physically prototyped so that I would not forget all the values etc.   This is the first version of that drawing - but I need to verify that it is correct by building a second prototype solely by referencing the schematic.